Tìm kiếm mã đọc giả | Tên độc giả
Tìm kiếm mã đọc giả | Tên độc giả (Test)

+++ admin/ajax/ajax_sua_nhan_vien.php
@@ -14,7 +14,7 @@
 			echo "<script type=\"text/javascript\">khongthanhcong(\"<strong>Chưa lưu</strong> HSPC không để trống!\")</script>";
 			exit();
 		}
-		settype($hspc, 'int');
+		settype($hspc, 'float');
 		$mk = md5("123456");
 		$ketnoi = new clsKetnoi();
 		$conn = $ketnoi->ketnoi();
