+++ admin/ajax/ajax_them_nhan_vien.php
@@ -26,7 +26,7 @@
 			echo "<script type=\"text/javascript\">khongthanhcong(\"<strong>Chưa lưu</strong> HSPC không để trống!\")</script>";
 			exit();
 		}
-		settype($hspc, 'int');
+		settype($hspc, 'float');
 		$mk = md5("123456");
 		$ketnoi = new clsKetnoi();
 		$conn = $ketnoi->ketnoi();
